requirements.txt
aiohttp==3.4.4
aiosqlite==0.8.1
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
discord.py==1.1.1
idna==2.8
multidict==4.5.2
pendulum==2.0.4
python-dateutil==2.8.0
pytzdata==2018.9
PyYAML==5.1.2
six==1.12.0
websockets==6.0
yarl==1.3.0
